+++ buybackprogram/views/programs.py
@@ -176,16 +176,19 @@ def location_add(request):
         if form.is_valid():
             eve_solar_system = form.cleaned_data["eve_solar_system"]
             name = form.cleaned_data["name"]
+            structure_id = form.cleaned_data["structure_id"]
 
             owner = Owner.objects.get(user=request.user)
 
             created = Location.objects.update_or_create(
                 eve_solar_system=eve_solar_system,
                 name=name,
+                structure_id=structure_id,
                 owner=owner,
                 defaults={
                     "eve_solar_system": eve_solar_system,
                     "name": name,
+                    "structure_id": structure_id,
                     "owner": owner,
                 },
             )