+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Fixed #10, added ability to delete own locations
 - Fixes #12, added requirement for eveuniverse in readme
+- Fixes #9, added ability to track contract locations per structure ID
 
 ### Changed
 - Fixes #11, now also tracking contracts that have extra characters in the description such as extra spaces.